Add "--copyright-year" command-line flag.
The --copyright-year=YYYY flag allows users to choose the year spec-cleaner
will use in the re-generated copyright header. This is useful for people who
want to have that attribution match the year of the latest entry in the
accompanying *.changes file. If unspecified, the value defaults to the current
year, like before.