Replace 'http' with 'https' in URL
Use 'https' in URL if a https website is reachable (issue #246). Also
update the old tests, add a new test (url_https.spec) for this feature
and mark this test as a webtest so we can ignore it while we don't
have an internet connection. Add docstrings for acceptance-tests.py.