Fix mingw32-clutter and replace_pwdutils tests
replace_pwdutils.spec and mingw32-clutter.spec tests contain URL
starting with http that could be replaced with https by spec-cleaner.
That however need internet connection that is not available during the
test phase in OBS. That cause that these 2 tests are failing there. We
could skip them by moving them to the web directory but that would be
a shame. I changed the http to https there so it's not faililing because of
that and we can still run the test for its original purpose (even in OBS).